Extract shared validation logic to prevent test divergence
